Pass big aggregated arguments indirectly.
Ensures that arguments of any type can be passed through
`classify_arg_ty()`. Previously any aggregated argument
with a size above the maximum of `MAX_ARG_IN_REGS_SIZE`
triggered the `stack_required` assertion.

The reason was that that the flag directly depends on the
calculated amount of required registered which is not
determined if a given argument is bigger than the expected
maximum of `MAX_ARG_IN_REGS_SIZE`.